resource-timing/entry-attributes.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing: PerformanceResourceTiming attributes</title>
<link rel="help" href="https://www.w3.org/TR/resource-timing-2/#sec-performanceresourcetiming"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

// Returns a promise that settles once the given path has been fetched.
function load_image(path) {
  return new Promise(resolve => {
    const img = new Image();
    img.onload = img.onerror = resolve;
    img.src = path;
  });
}

// Returns the PerformanceResourceTiming entry that matches the given
// predicate. If there isn't exactly one matching entry, an error is thrown.
function find_entry(predicate) {
  const entries = performance.getEntriesByType('resource').filter(predicate);
  if (entries.length != 1) {
    throw new Error("There should be one entry");
  }
  return entries[0];
}

promise_test(async () => {
  await load_image("resources/fake_responses.py#hash=1");
  const entry = find_entry(e => e.initiatorType === 'img');
  assert_true(entry.name.includes('#hash=1'),
      "There should be a hash in the resource name");
}, "Url fragments should be present in the 'name' attribute");

</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that PerformanceResourceTiming entries' attributes are
populated with the correct values.</p>
</body>
</html>
